Backend/requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
pydantic-settings
sqlalchemy
pymysql
cryptography